Move has_converged to util
